pyproject.toml
[project]
name = "gov-bizmon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
]
